split tests into two files
